+++ client/java/src/main/java/org/jolokia/client/request/ValidatingResponseExtractor.java
@@ -19,7 +19,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 import org.jolokia.client.exception.J4pRemoteException;
-import org.json.simple.JSONObject;
+import org.jolokia.json.JSONObject;
 
 /**
  * A response extractor which does validation based on Jolokia status codes.
@@ -58,8 +58,8 @@ public class ValidatingResponseExtractor implements J4pResponseExtractor {
         int status = 0;
         if (pJsonResp.containsKey("status")) {
             Object o = pJsonResp.get("status");
-            if (o instanceof Long) {
-                status = ((Long) o).intValue();
+            if (o instanceof Integer) {
+                status = (Integer) o;
             }
         }
